doc/script/pyproject.toml
[project]
name = "script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prompt-toolkit>=3.0.52",
    "pyserial>=3.5",
]


[tool.ruff.lint]
ignore = ["E501"]
extend-select = ["I"]
